+++ attachment-viewer/src/main/java/im/vector/lib/attachmentviewer/AnimatedImageViewHolder.kt
@@ -14,18 +14,17 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.attachmentviewer
+package im.vector.lib.attachmentviewer
 
-sealed class AttachmentEvents {
-    data class VideoEvent(val isPlaying: Boolean, val progress: Int, val duration: Int) : AttachmentEvents()
-}
+import android.view.View
+import android.widget.ImageView
+import android.widget.ProgressBar
 
-interface AttachmentEventListener {
-    fun onEvent(event: AttachmentEvents)
-}
+class AnimatedImageViewHolder constructor(itemView: View) :
+        BaseViewHolder(itemView) {
 
-sealed class AttachmentCommands {
-    object PauseVideo : AttachmentCommands()
-    object StartVideo : AttachmentCommands()
-    data class SeekTo(val percentProgress: Int) : AttachmentCommands()
+    val touchImageView: ImageView = itemView.findViewById(R.id.imageView)
+    val imageLoaderProgress: ProgressBar = itemView.findViewById(R.id.imageLoaderProgress)
+
+    internal val target = DefaultImageLoaderTarget(this, this.touchImageView)
 }