+++ attachment-viewer/src/main/java/im/vector/lib/attachmentviewer/AttachmentViewerActivity.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.attachmentviewer
+package im.vector.lib.attachmentviewer
 
 import android.graphics.Color
 import android.os.Bundle