+++ attachment-viewer/src/main/java/im/vector/lib/attachmentviewer/AttachmentsAdapter.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.attachmentviewer
+package im.vector.lib.attachmentviewer
 
 import android.view.LayoutInflater
 import android.view.ViewGroup