+++ attachment-viewer/src/main/java/im/vector/lib/attachmentviewer/SwipeDirection.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.attachmentviewer
+package im.vector.lib.attachmentviewer
 
 sealed class SwipeDirection {
     object NotDetected : SwipeDirection()