+++ attachment-viewer/src/main/java/im/vector/lib/attachmentviewer/SwipeDirectionDetector.kt
@@ -1,6 +1,5 @@
 /*
  * Copyright (c) 2020 New Vector Ltd
- * Copyright (C) 2018 stfalcon.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.attachmentviewer
+package im.vector.lib.attachmentviewer
 
 import android.content.Context
 import android.view.MotionEvent